+++ tools/analysis/__init__.py
@@ -1,0 +1,7 @@
+"""Code analysis tools (type checking, linting)."""
+
+from tools.analysis.check_types import CheckTypesTool
+
+__all__ = [
+    "CheckTypesTool",
+]